requirements.txt
astroid==2.2.0
colorama==0.4.1
Django==2.1.7
djangorestframework==3.9.1
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pylint==2.3.0
pytz==2018.9
six==1.12.0
sorl-thumbnail==12.5.0
typed-ast==1.3.1
wrapt==1.11.1
